Fix lossy conversion errors; note that we'll have to scale the original values in our adapted controller

+++ Communications/sdp/sdp9/strategy/PassingStrategy.java
@@ -561,7 +561,7 @@ public class PassingStrategy extends GeneralStrategy {
 						attackerBrick.execute(new RobotCommand.Travel(
 								travelDist, travelSpeed));
 						defenderBrick.execute(new RobotCommand.Rotate(rotateBy,
-								Math.abs(rotateBy)));
+								(byte)Math.abs(rotateBy)));
 						break;
 					case MOVENROTATE:
 						defenderBrick.executeSync(new RobotCommand.Travel(
